antigravity/requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
Django>=6.0.1

# ── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary          # PostgreSQL adapter

# ── Secrets management ───────────────────────────────────────────────────────
python-decouple          # Read secrets from .env — never hardcode them

# ── Security packages ────────────────────────────────────────────────────────
django-axes              # Brute-force login protection (auto-lockout)
django-ratelimit         # Per-view rate limiting by IP
django-csp               # Content Security Policy response headers
django-cors-headers      # CORS control for API endpoints
whitenoise               # Secure static file serving without a separate web server

# ── Password hashing ─────────────────────────────────────────────────────────
bcrypt                   # Stronger password hashing backend

# ── Dev/CI tools (not needed in production image) ────────────────────────────
bandit                   # Static security linting — catches known Python vulns
pre-commit               # Git hook manager — runs checks before every commit
